Add files via upload
TicTacToe game, followed a tutorial, and re-made it over and over until i understood it! i focused on what was going on step by step very analytically to study how the functions work, passing parameters and looping through arrays to check if the set entered by the user contains a winning set of numbers in the arrays, etc.